pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cade"
version = "0.1.0"
description = "Counterparty Agreement Data Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pyarrow>=15.0",
    "pandas>=2.2",
    "typer[all]>=0.12",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "boto3>=1.26",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "moto[s3]>=5.0",
    "testcontainers[postgres]>=4.0",
]

[project.scripts]
cade = "cade.cli:app"
cade-ingest = "cade.ingest.runner:app"

[tool.setuptools.packages.find]
include = ["cade*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
